nrfx: gppi: fix to use only tasks and events as endpoints
So far endpoints could accept publish/subscribe registers and did
not use hal to control them. There is no use case for publish/subscribe
as input arguments so reworked GPPI to accept only tasks and events
register addresses. HAL is used to control them. HAL needs to be used
to work with babblesim.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>